agent/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
httpx==0.27.0
anthropic==0.28.0
langchain==0.2.1
langchain-anthropic==0.1.13
langgraph==0.1.1
redis==5.0.4
python-dotenv==1.0.1
web3==6.18.0
pytest==8.2.2
